fix(endgame conditions): Initially saw this as a bug, but it's not!
This could be one of the key mechanics - that seemingly impossible
comebacks are available, if one of the players run out of moves - their
opponent always wins, no matter the scores.

This is very harsh, but helps the underdog and gives them something to
fight for.

Added a map to test this on.